--- .npmignore
@@ -1,3 +1,0 @@
-.git*
-test/
-docs
